refactor sprint package structure
