+++ backend/src/main/java/com/tissue/api/issue/application/service/finder/IssueFinder.java
@@ -11,7 +11,7 @@ import com.tissue.api.issue.domain.Issue;
 import com.tissue.api.issue.domain.enums.StateCategory;
 import com.tissue.api.issue.domain.exception.IssueNotFoundException;
 import com.tissue.api.issue.domain.port.out.IssueQueryRepository;
-import com.tissue.api.sprint.domain.model.Sprint;
+import com.tissue.api.sprint.domain.Sprint;
 
 import lombok.RequiredArgsConstructor;
