+++ backend/src/main/java/com/tissue/api/security/authorization/SprintSecurityGuard.java
@@ -2,9 +2,9 @@ package com.tissue.api.security.authorization;
 
 import org.springframework.stereotype.Component;
 
-import com.tissue.api.sprint.domain.model.Sprint;
-import com.tissue.api.sprint.exception.SprintNotFoundException;
-import com.tissue.api.sprint.infrastructure.repository.SprintQueryRepository;
+import com.tissue.api.sprint.domain.Sprint;
+import com.tissue.api.sprint.domain.exception.SprintNotFoundException;
+import com.tissue.api.sprint.domain.port.out.SprintQueryRepository;
 
 import lombok.RequiredArgsConstructor;
