+++ backend/src/main/java/com/tissue/api/sprint/application/dto/response/SprintDetail.java
@@ -2,8 +2,8 @@ package com.tissue.api.sprint.application.dto.response;
 
 import java.time.Instant;
 
-import com.tissue.api.sprint.domain.model.Sprint;
-import com.tissue.api.sprint.domain.model.enums.SprintStatus;
+import com.tissue.api.sprint.domain.Sprint;
+import com.tissue.api.sprint.domain.enums.SprintStatus;
 
 import lombok.Builder;
