+++ backend/src/main/java/com/tissue/api/sprint/application/service/finder/SprintFinder.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 
 import com.tissue.api.project.domain.Project;
-import com.tissue.api.sprint.domain.model.Sprint;
-import com.tissue.api.sprint.domain.model.enums.SprintStatus;
-import com.tissue.api.sprint.exception.SprintNotFoundException;
-import com.tissue.api.sprint.infrastructure.repository.SprintQueryRepository;
+import com.tissue.api.sprint.domain.Sprint;
+import com.tissue.api.sprint.domain.enums.SprintStatus;
+import com.tissue.api.sprint.domain.exception.SprintNotFoundException;
+import com.tissue.api.sprint.domain.port.out.SprintQueryRepository;
 
 import lombok.RequiredArgsConstructor;
