+++ backend/src/main/java/com/tissue/api/sprint/domain/enums/SprintStatus.java
@@ -1,4 +1,4 @@
-package com.tissue.api.sprint.domain.model.enums;
+package com.tissue.api.sprint.domain.enums;
 
 public enum SprintStatus {
 	PLANNING,
